Fix Bug 88: Changing selected instances does not remove old ones from topics (but no listening on them)

+++ JERunner/RuleEngine/src/main/java/io/je/ruleengine/impl/ProjectContainer.java
@@ -113,6 +113,16 @@ public class ProjectContainer {
 
     }
 
+    void logError(Exception exp, String message) {
+        logError(exp, message, null);
+    }
+
+    void logError(Exception exp, String message, String objectId) {
+        exp.printStackTrace();
+        JELogger.debugWithoutPublish(Arrays.toString(exp.getStackTrace()),LogCategory.RUNTIME,projectId, LogSubModule.RULE,objectId);
+        JELogger.error(message, LogCategory.RUNTIME, projectId, LogSubModule.RULE,objectId);
+    }
+
     public void resetContainer() {
         JELogger.debug(JEMessages.RELOADING_PROJECT_CONTAINER, LogCategory.RUNTIME, projectId, LogSubModule.RULE,
                 null);
@@ -132,14 +142,13 @@ public class ProjectContainer {
         stopRuleExecution(true, true);
         try {
             fireRules();
-        } catch (RulesNotFiredException | RuleBuildFailedException e) {
-            e.printStackTrace();
-            JELogger.error(JEMessages.FAILED_TO_FIRE_RULES, LogCategory.RUNTIME, projectId, LogSubModule.RULE,
-                    null);
+        } catch (RulesNotFiredException | RuleBuildFailedException exp) {
+            logError(exp, JEMessages.FAILED_TO_FIRE_RULES);
         }
         reloadContainer = false;
         JELogger.trace("Reloaded Rule Engine.");
     }
+
 
     /*--------------------------------------------------------PROJECT METHODS ---------------------------------------------------------------*/
 
@@ -158,6 +167,7 @@ public class ProjectContainer {
         JELogger.debugWithoutPublish(JEMessages.BUILDING_PROJECT_CONTAINER, LogCategory.RUNTIME, projectId,
                 LogSubModule.RULE, null);
         // loader = new JEClassLoader(ProjectContainer.class.getClassLoader());
+
         // build kie environment
         if (!buildKie()) {
             JELogger.error(JEMessages.BUILDING_PROJECT_CONTAINER_FAILED, LogCategory.RUNTIME, projectId,
@@ -202,9 +212,9 @@ public class ProjectContainer {
             }
             Runnable runnable = () -> {
                 try {
-                    //https://docs.drools.org/7.70.0.Final/drools-docs/html_single/index.html#_event_model
-                    kieSession.addEventListener(new DebugAgendaEventListener());
-                    kieSession.addEventListener(new DebugRuleRuntimeEventListener());
+                    // https://docs.drools.org/7.70.0.Final/drools-docs/html_single/index.html#_event_model
+                    // kieSession.addEventListener(new DebugAgendaEventListener());
+                    // kieSession.addEventListener(new DebugRuleRuntimeEventListener());
 
                     //  kieSession.addEventListener(new RuleListener(projectId));
                     //  kieSession.addEventListener(ruleListener);
@@ -217,14 +227,12 @@ public class ProjectContainer {
                     // fatal : Runtime Executions
                     String ruleId = IdManager.getRuleIdFromErrorMsg(exp.getMessage());
 
-                    JELogger.error(
-                            JEMessages.RULE_EXECUTION_ERROR + StringUtils.substringBefore(exp.getMessage(), " in "),
-                            LogCategory.RUNTIME, projectId, LogSubModule.RULE, ruleId);
+                    logError(exp, JEMessages.RULE_EXECUTION_ERROR + StringUtils.substringBefore(exp.getMessage(), " in "), ruleId);
+
                     try {
                         fireRules();
-                    } catch (RulesNotFiredException | RuleBuildFailedException e1) {
-                        JELogger.error(JEMessages.FAILED_TO_FIRE_RULES, LogCategory.RUNTIME, projectId,
-                                LogSubModule.RULE, ruleId);
+                    } catch (RulesNotFiredException | RuleBuildFailedException exp1) {
+                        logError(exp1, JEMessages.FAILED_TO_FIRE_RULES);
                     }
 
                 }
@@ -237,8 +245,8 @@ public class ProjectContainer {
             if (t1 != null) {
                 kieSession.halt();
             }
-            JELogger.error(JEMessages.FAILED_TO_FIRE_RULES + " : " + exp.getMessage(), LogCategory.RUNTIME, projectId,
-                    LogSubModule.RULE, null);
+
+            logError(exp, JEMessages.FAILED_TO_FIRE_RULES);
 
             throw new RulesNotFiredException(Arrays.toString(exp.getStackTrace()));
 
@@ -271,10 +279,8 @@ public class ProjectContainer {
 
             }
 
-        } catch (Exception e) {
-            e.printStackTrace();
-            JELogger.error(JEMessages.STOPPING_PROJECT_CONTAINER_FAILED, LogCategory.RUNTIME, projectId,
-                    LogSubModule.RULE, null);
+        } catch (Exception exp) {
+            logError(exp, JEMessages.STOPPING_PROJECT_CONTAINER_FAILED);
             return false;
         }
         return true;
@@ -308,10 +314,8 @@ public class ProjectContainer {
         try {
             kieServices.newKieBuilder(kieFileSystem, JEClassLoader.getDataModelInstance())
                     .buildAll(null);
-        } catch (Exception e) {
-            // e.printStackTrace();
-            JELogger.error("Error creating kieBuilder \n " + Arrays.toString(e.getStackTrace()), LogCategory.RUNTIME,
-                    projectId, LogSubModule.RULE, null);
+        } catch (Exception exp) {
+            logError(exp, "Error creating kieBuilder");
             return false;
         }
 
@@ -338,9 +342,8 @@ public class ProjectContainer {
                 Thread.currentThread()
                         .setContextClassLoader(JEClassLoader.getDataModelInstance());
 
-            } catch (Exception e) {
-                JELogger.error("Error creating kieBase \n " + Arrays.toString(e.getStackTrace()),
-                        LogCategory.RUNTIME, projectId, LogSubModule.RULE, null);
+            } catch (Exception exp) {
+                logError(exp, "Error creating kieBase");
                 return false;
             }
 
@@ -349,7 +352,6 @@ public class ProjectContainer {
         }
         return true;
     }
-
 
     /*
      * this method is responsible for creating the kieModule
@@ -379,9 +381,8 @@ public class ProjectContainer {
 
             // generate pom file
             kieFileSystem.generateAndWritePomXML(releaseId);
-        } catch (Exception e) {
-            JELogger.error(JEMessages.UNEXPECTED_ERROR + "\n" + Arrays.toString(e.getStackTrace()), LogCategory.RUNTIME,
-                    projectId, LogSubModule.RULE, null);
+        } catch (Exception exp) {
+            logError(exp, JEMessages.UNEXPECTED_ERROR);
         }
 
     }
@@ -404,13 +405,17 @@ public class ProjectContainer {
      */
     private boolean addRuleToKieFileSystem(Rule rule) {
 
+        if (rule == null || rule.getJobEngineElementName() == null) return false;
+
         try {
             String ruleName = generateResourceName(ResourceType.DRL, rule.getJobEngineElementName());
-            JELogger.trace(">>> adding " + ruleName, LogCategory.RUNTIME, projectId, LogSubModule.RULE, null);
+
+            JELogger.trace(">>> adding ", LogCategory.RUNTIME, projectId, LogSubModule.RULE, ruleName);
+
             kieFileSystem.write(ruleName, rule.getContent());
 
         } catch (Exception exp) {
-            JELogger.error(JEMessages.FAILED_TO_ADD_RULE, LogCategory.RUNTIME, projectId, LogSubModule.RULE, null);
+            logError(exp, JEMessages.FAILED_TO_ADD_RULE);
             return false;
         }
 
@@ -422,16 +427,19 @@ public class ProjectContainer {
      */
     private boolean deleteRuleFromKieFileSystem(Rule rule) {
 
+        if (rule == null || rule.getJobEngineElementName() == null) return false;
+
         try {
             String ruleName = generateResourceName(ResourceType.DRL, rule.getJobEngineElementName());
-            JELogger.trace("deleting " + ruleName, LogCategory.RUNTIME, projectId, LogSubModule.RULE, null);
+
+            JELogger.trace(">>> deleting ", LogCategory.RUNTIME, projectId, LogSubModule.RULE, ruleName);
 
             if (kieFileSystem.read(ruleName) != null) {
                 kieFileSystem.delete(ruleName);
             }
 
-        } catch (Exception e) {
-            JELogger.error(JEMessages.FAILED_TO_DELETE_RULE, LogCategory.RUNTIME, projectId, LogSubModule.RULE, null);
+        } catch (Exception exp) {
+            logError(exp, JEMessages.FAILED_TO_DELETE_RULE, rule.getJobEngineElementName());
             return false;
         }
 
@@ -476,8 +484,6 @@ public class ProjectContainer {
             }
             // Thread.currentThread().setContextClassLoader(JEClassLoader.getInstance());
 
-            //kieContainer.updateToVersion(releaseId);
-            //startTime = System.nanoTime();
             kieContainer.updateToVersion(releaseId);
 
             if (kScanner == null) {
@@ -485,13 +491,9 @@ public class ProjectContainer {
             }
             //kScanner.scanNow(); removed because it consults the internet
 
-        } catch (Exception e) {
-            // FIXME case of error should return false?
-            e.printStackTrace();
-            JELogger.error(JEMessages.UNEXPECTED_ERROR + "\n" + e.getMessage(), LogCategory.RUNTIME, projectId,
-                    LogSubModule.RULE, null);
-            // JELogger.error(ProjectContainer.class,
-            // RuleEngineErrors.FailedToUpdateContainer + e.getMessage());
+        } catch (Exception exp) {
+            logError(exp, JEMessages.UNEXPECTED_ERROR);
+            return false;
         }
 
         return true;
@@ -526,6 +528,7 @@ public class ProjectContainer {
         JELogger.debugWithoutPublish("[projectId =" + projectId + "] " + JEMessages.ADDING_RULE + " ["
                         + rule.getJobEngineElementName() + "]", LogCategory.RUNTIME, projectId, LogSubModule.RULE,
                 rule.getJobEngineElementID());
+
         // check if rule already exists
         if (ruleExists(rule)) {
             throw new RuleAlreadyExistsException(JEMessages.RULE_EXISTS);
@@ -533,9 +536,11 @@ public class ProjectContainer {
 
         // compile rule
         compileRule(rule);
-        allRules.put(rule.getJobEngineElementID(), rule);
-        // FIXME check return value
-        addRuleToKieFileSystem(rule);
+
+        if (addRuleToKieFileSystem(rule)) {
+            allRules.put(rule.getJobEngineElementID(), rule);
+        }
+
         updateContainer();
 
         // if project is running
@@ -549,35 +554,34 @@ public class ProjectContainer {
      * rule exists => rule will be updates
      */
     public boolean updateRule(Rule rule) throws RuleCompilationException, JEFileNotFoundException {
-        JELogger.debugWithoutPublish("[projectId =" + projectId + "] " + JEMessages.UPDATING_RULE + " ["
-                        + rule.getJobEngineElementName() + "]", LogCategory.RUNTIME, projectId, LogSubModule.RULE,
-                rule.getJobEngineElementID());
-        // compile rule
-        compileRule(rule);
-
-        if (status != Status.RUNNING) {
-            buildStatus = BuildStatus.UNBUILT;
-        }
-
-        // check that rule exists and add it if not
-        if (!ruleExists(rule)) {
-            allRules.put(rule.getJobEngineElementID(), rule);
-            if (!addRuleToKieFileSystem(rule)) { return false; };
-            updateContainer();
-            return true;
-        }
-
-        // update rule in map
-        allRules.put(rule.getJobEngineElementID(), rule);
-
-        // if project is running, update container without interrupting project
         try {
+            JELogger.debugWithoutPublish("[projectId =" + projectId + "] " + JEMessages.UPDATING_RULE + " ["
+                            + rule.getJobEngineElementName() + "]", LogCategory.RUNTIME, projectId, LogSubModule.RULE,
+                    rule.getJobEngineElementID());
+            // compile rule
+            compileRule(rule);
+
+            if (status != Status.RUNNING) {
+                buildStatus = BuildStatus.UNBUILT;
+            }
+
+            // check that rule exists and add it if not
+            if (!ruleExists(rule)) {
+                allRules.put(rule.getJobEngineElementID(), rule);
+                if (!addRuleToKieFileSystem(rule)) { return false; };
+                updateContainer();
+                return true;
+            }
+
+            // update rule in map
+            allRules.put(rule.getJobEngineElementID(), rule);
+
+            // if project is running, update container without interrupting project
             if (!deleteRuleFromKieFileSystem(rule)) { return false; };
             if (!addRuleToKieFileSystem(rule)) { return false; };
             updateContainer();
-        } catch (Exception e) {
-            JELogger.error(JEMessages.RULE_UPDATE_FAIL + "\n" + Arrays.toString(e.getStackTrace()), LogCategory.RUNTIME,
-                    projectId, LogSubModule.RULE, rule.getJobEngineElementID());
+        } catch (Exception exp) {
+            logError(exp, JEMessages.RULE_UPDATE_FAIL, rule.getJobEngineElementID());
             return false;
         }
 
@@ -589,33 +593,35 @@ public class ProjectContainer {
      * delete rule from engine
      */
     public void deleteRule(String ruleId) throws DeleteRuleException {
-        JELogger.debugWithoutPublish(
-                "[projectId = " + projectId + "] [ruleId : " + ruleId + "] " + JEMessages.DELETING_RULE ,
-                LogCategory.RUNTIME, projectId, LogSubModule.RULE, ruleId);
-        // check that rule exists
-        if (!ruleExists(ruleId)) {
-            return;
-        }
-        // if project is running, update container without interrupting project
         try {
+            JELogger.debugWithoutPublish(
+                    "[projectId = " + projectId + "] [ruleId : " + ruleId + "] " + JEMessages.DELETING_RULE ,
+                    LogCategory.RUNTIME, projectId, LogSubModule.RULE, ruleId);
+            // check that rule exists
+            if (!ruleExists(ruleId)) {
+                return;
+            }
+            // if project is running, update container without interrupting project
             long startTime = System.nanoTime();
-            // FIXME check return value
+
+            // FIXME check returned value
             deleteRuleFromKieFileSystem(allRules.get(ruleId));
             long endTime = System.nanoTime();
             long duration = (endTime - startTime) / 1000000; //divide by 1000000 to get milliseconds.
             JELogger.debug("deleteRuleFromKieFileSystem : " + duration);
 
             updateContainer();
-        } catch (Exception e) {
-            JELogger.error(JEMessages.RULE_DELETE_FAIL + "\n" + Arrays.toString(e.getStackTrace()), LogCategory.RUNTIME,
-                    projectId, LogSubModule.RULE, ruleId);
+
+        } catch (Exception exp) {
+
+            logError(exp, JEMessages.RULE_DELETE_FAIL, ruleId);
+
             throw new DeleteRuleException(JEMessages.RULE_DELETE_FAIL);
 
         }
         allRules.remove(ruleId);
 
         if (status != Status.RUNNING) {
-
             buildStatus = BuildStatus.UNBUILT;
         }
 
@@ -651,8 +657,9 @@ public class ProjectContainer {
                 .buildAll(null);
 
         Results results = kieBuilder.getResults();
-        //TODO to check utility
+        //FIXME to check utility
         kfsToCompile.delete(filename);
+
         if (results.hasMessages(Message.Level.ERROR)) {
             JELogger.error(results.getMessages()
                             .toString(), LogCategory.RUNTIME, projectId, LogSubModule.RULE,
@@ -661,7 +668,6 @@ public class ProjectContainer {
                     .get(0)
                     .getText());
         }
-
 
     }
 
@@ -769,12 +775,11 @@ public class ProjectContainer {
 
                         }
                     }
-                } catch (Exception e) {
-                    e.printStackTrace();
-                    JELogger.error(
-                            "[factId =" + fact.getJobEngineElementID() + "]" + JEMessages.FAILED_TO_UPDATE_FACT + ": "
-                                    + e.getMessage() + "\n",
-                            LogCategory.RUNTIME, projectId, LogSubModule.RULE, fact.getJobEngineElementID());
+
+                } catch (Exception exp) {
+
+                    logError(exp, "[factId =" + fact.getJobEngineElementID() + "] " + JEMessages.FAILED_TO_UPDATE_FACT,
+                            fact.getJobEngineElementID());
 
                 }
 
@@ -812,9 +817,6 @@ public class ProjectContainer {
     public void setReloadContainer(boolean reloadContainer) {
         this.reloadContainer = reloadContainer;
     }
-
-
-
 
     /*
      * public void setClassLoader(JEClassLoader loader) { this.loader = loader;
